convert more files to typescript
