+++ package.json
@@ -80,8 +80,9 @@
     "eslint-plugin-unicorn": "^49.0.0",
     "gh-pages": "^4.0.0",
     "prettier": "^3.0.3",
-    "typescript": "^4.0.8",
-    "typescript-eslint": "^0.0.1-alpha.0"
+    "typescript": "4.8.4",
+    "typescript-eslint": "^0.0.1-alpha.0",
+    "typescript-plugin-styled-components": "^3.0.0"
   },
   "babel": {
     "presets": [
